+++ elliptic-curve/src/hash2field/expand_msg/xmd.rs
@@ -1,33 +1,108 @@
 //! `expand_message_xmd` based on a hash function.
 
-use super::{Domain, ExpandMsg};
+use core::marker::PhantomData;
+
+use super::{Domain, ExpandMsg, Expander};
 use crate::{Error, Result};
 use digest::{
+    core_api::BlockSizeUser,
     generic_array::{
         typenum::{IsLess, IsLessOrEqual, Unsigned, U256},
         GenericArray,
     },
-    BlockInput, Digest,
+    Digest,
 };
 
 /// Placeholder type for implementing `expand_message_xmd` based on a hash function
-pub struct ExpandMsgXmd<HashT>
+///
+/// # Errors
+/// - `len_in_bytes == 0`
+/// - `len_in_bytes > u16::MAX`
+/// - `len_in_bytes > 255 * HashT::OutputSize`
+pub struct ExpandMsgXmd<HashT>(PhantomData<HashT>)
 where
-    HashT: Digest + BlockInput,
+    HashT: Digest + BlockSizeUser,
+    HashT::OutputSize: IsLess<U256>,
+    HashT::OutputSize: IsLessOrEqual<HashT::BlockSize>;
+
+/// ExpandMsgXmd implements expand_message_xmd for the ExpandMsg trait
+impl<'a, HashT> ExpandMsg<'a> for ExpandMsgXmd<HashT>
+where
+    HashT: Digest + BlockSizeUser,
+    // If `len_in_bytes` is bigger then 256, length of the `DST` will depend on
+    // the output size of the hash, which is still not allowed to be bigger then 256:
+    // https://www.ietf.org/archive/id/draft-irtf-cfrg-hash-to-curve-13.html#section-5.4.1-6
+    HashT::OutputSize: IsLess<U256>,
+    // Constraint set by `expand_message_xmd`:
+    // https://www.ietf.org/archive/id/draft-irtf-cfrg-hash-to-curve-13.html#section-5.4.1-4
+    HashT::OutputSize: IsLessOrEqual<HashT::BlockSize>,
+{
+    type Expander = ExpanderXmd<'a, HashT>;
+
+    fn expand_message(
+        msgs: &[&[u8]],
+        dst: &'a [u8],
+        len_in_bytes: usize,
+    ) -> Result<Self::Expander> {
+        if len_in_bytes == 0 {
+            return Err(Error);
+        }
+
+        let len_in_bytes_u16 = u16::try_from(len_in_bytes).map_err(|_| Error)?;
+
+        let b_in_bytes = HashT::OutputSize::to_usize();
+        let ell = u8::try_from((len_in_bytes + b_in_bytes - 1) / b_in_bytes).map_err(|_| Error)?;
+
+        let domain = Domain::xmd::<HashT>(dst);
+        let mut b_0 = HashT::new();
+        b_0.update(GenericArray::<u8, HashT::BlockSize>::default());
+
+        for msg in msgs {
+            b_0.update(msg);
+        }
+
+        b_0.update(len_in_bytes_u16.to_be_bytes());
+        b_0.update([0]);
+        b_0.update(domain.data());
+        b_0.update([domain.len()]);
+        let b_0 = b_0.finalize();
+
+        let mut b_vals = HashT::new();
+        b_vals.update(&b_0[..]);
+        b_vals.update([1u8]);
+        b_vals.update(domain.data());
+        b_vals.update([domain.len()]);
+        let b_vals = b_vals.finalize();
+
+        Ok(ExpanderXmd {
+            b_0,
+            b_vals,
+            domain,
+            index: 1,
+            offset: 0,
+            ell,
+        })
+    }
+}
+
+/// [`Expander`] type for [`ExpandMsgXmd`].
+pub struct ExpanderXmd<'a, HashT>
+where
+    HashT: Digest + BlockSizeUser,
     HashT::OutputSize: IsLess<U256>,
     HashT::OutputSize: IsLessOrEqual<HashT::BlockSize>,
 {
     b_0: GenericArray<u8, HashT::OutputSize>,
     b_vals: GenericArray<u8, HashT::OutputSize>,
-    domain: Domain<HashT::OutputSize>,
+    domain: Domain<'a, HashT::OutputSize>,
     index: u8,
     offset: usize,
     ell: u8,
 }
 
-impl<HashT> ExpandMsgXmd<HashT>
+impl<'a, HashT> ExpanderXmd<'a, HashT>
 where
-    HashT: Digest + BlockInput,
+    HashT: Digest + BlockSizeUser,
     HashT::OutputSize: IsLess<U256>,
     HashT::OutputSize: IsLessOrEqual<HashT::BlockSize>,
 {
@@ -42,12 +117,12 @@ where
                 .zip(&self.b_vals[..])
                 .enumerate()
                 .for_each(|(j, (b0val, bi1val))| tmp[j] = b0val ^ bi1val);
-            self.b_vals = HashT::new()
-                .chain(tmp)
-                .chain([self.index])
-                .chain(self.domain.data())
-                .chain([self.domain.len()])
-                .finalize();
+            let mut b_vals = HashT::new();
+            b_vals.update(tmp);
+            b_vals.update([self.index]);
+            b_vals.update(self.domain.data());
+            b_vals.update([self.domain.len()]);
+            self.b_vals = b_vals.finalize();
             true
         } else {
             false
@@ -55,55 +130,12 @@ where
     }
 }
 
-/// ExpandMsgXmd implements expand_message_xmd for the ExpandMsg trait
-impl<HashT> ExpandMsg for ExpandMsgXmd<HashT>
+impl<'a, HashT> Expander for ExpanderXmd<'a, HashT>
 where
-    HashT: Digest + BlockInput,
-    // If `len_in_bytes` is bigger then 256, length of the `DST` will depend on
-    // the output size of the hash, which is still not allowed to be bigger then 256:
-    // https://www.ietf.org/archive/id/draft-irtf-cfrg-hash-to-curve-13.html#section-5.4.1-6
+    HashT: Digest + BlockSizeUser,
     HashT::OutputSize: IsLess<U256>,
-    // Constraint set by `expand_message_xmd`:
-    // https://www.ietf.org/archive/id/draft-irtf-cfrg-hash-to-curve-13.html#section-5.4.1-4
     HashT::OutputSize: IsLessOrEqual<HashT::BlockSize>,
 {
-    fn expand_message(msg: &[u8], dst: &'static [u8], len_in_bytes: usize) -> Result<Self> {
-        if len_in_bytes == 0 {
-            return Err(Error);
-        }
-
-        let len_in_bytes_u16 = u16::try_from(len_in_bytes).map_err(|_| Error)?;
-
-        let b_in_bytes = HashT::OutputSize::to_usize();
-        let ell = u8::try_from((len_in_bytes + b_in_bytes - 1) / b_in_bytes).map_err(|_| Error)?;
-
-        let domain = Domain::xmd::<HashT>(dst);
-        let b_0 = HashT::new()
-            .chain(GenericArray::<u8, HashT::BlockSize>::default())
-            .chain(msg)
-            .chain(len_in_bytes_u16.to_be_bytes())
-            .chain([0])
-            .chain(domain.data())
-            .chain([domain.len()])
-            .finalize();
-
-        let b_vals = HashT::new()
-            .chain(&b_0[..])
-            .chain([1u8])
-            .chain(domain.data())
-            .chain([domain.len()])
-            .finalize();
-
-        Ok(Self {
-            b_0,
-            b_vals,
-            domain,
-            index: 1,
-            offset: 0,
-            ell,
-        })
-    }
-
     fn fill_bytes(&mut self, okm: &mut [u8]) {
         for b in okm {
             if self.offset == self.b_vals.len() && !self.next() {
@@ -128,11 +160,11 @@ mod test {
 
     fn assert_message<HashT>(
         msg: &[u8],
-        domain: &Domain<HashT::OutputSize>,
+        domain: &Domain<'_, HashT::OutputSize>,
         len_in_bytes: u16,
         bytes: &[u8],
     ) where
-        HashT: Digest + BlockInput,
+        HashT: Digest + BlockSizeUser,
         HashT::OutputSize: IsLess<U256>,
     {
         let block = HashT::BlockSize::to_usize();
@@ -169,16 +201,16 @@ mod test {
         fn assert<HashT, L: ArrayLength<u8>>(
             &self,
             dst: &'static [u8],
-            domain: &Domain<HashT::OutputSize>,
+            domain: &Domain<'_, HashT::OutputSize>,
         ) -> Result<()>
         where
-            HashT: Digest + BlockInput,
+            HashT: Digest + BlockSizeUser,
             HashT::OutputSize: IsLess<U256> + IsLessOrEqual<HashT::BlockSize>,
         {
             assert_message::<HashT>(self.msg, domain, L::to_u16(), self.msg_prime);
 
             let mut expander =
-                <ExpandMsgXmd<HashT> as ExpandMsg>::expand_message(self.msg, dst, L::to_usize())?;
+                ExpandMsgXmd::<HashT>::expand_message(&[self.msg], dst, L::to_usize())?;
 
             let mut uniform_bytes = GenericArray::<u8, L>::default();
             expander.fill_bytes(&mut uniform_bytes);